#4811-alternative: alternative approach

+++ factcast-store/src/main/java/org/factcast/store/internal/catchup/chunked/PgChunkedCatchup.java
@@ -30,7 +30,6 @@ import org.factcast.store.internal.*;
 import org.factcast.store.internal.catchup.*;
 import org.factcast.store.internal.pipeline.ServerPipeline;
 import org.factcast.store.internal.pipeline.Signal;
-import org.factcast.store.internal.query.CurrentStatementHolder;
 import org.factcast.store.internal.query.PgQueryBuilder;
 import org.factcast.store.internal.rowmapper.PgFactExtractor;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -49,10 +48,9 @@ public class PgChunkedCatchup extends AbstractPgCatchup {
       @NonNull SubscriptionRequestTO req,
       @NonNull ServerPipeline pipeline,
       @NonNull AtomicLong serial,
-      @NonNull CurrentStatementHolder statementHolder,
       @NonNull SingleConnectionDataSource ds,
       PgCatchupFactory.@NonNull Phase phase) {
-    super(props, metrics, req, pipeline, serial, statementHolder, ds, phase);
+    super(props, metrics, req, pipeline, serial, ds, phase);
   }
 
   @SneakyThrows
@@ -74,53 +72,50 @@ public class PgChunkedCatchup extends AbstractPgCatchup {
 
     String tempTableName = "catchup_" + UUID.randomUUID().toString().replace("-", "");
 
+    // this needs to be transactional for fetch-size to have any effect whatsoever.
+    Connection connection = ds.getConnection();
+    connection.setAutoCommit(false);
+    jdbc.setFetchSize(props.getPageSize());
+    jdbc.setQueryTimeout(0); // disable query timeout
+    if (prepareTemporaryTable(jdbc, tempTableName) > 0) {
+
+      if (wasCancelled()) return;
+
+      final var extractor = new PgFactExtractor(serial);
+
+      String chunkQuery = prepareChunkQuery(tempTableName);
+
+      int chunkCount = 0;
+      int rowsToProcess = -1;
+      while (rowsToProcess != 0) {
+
+        if (wasCancelled()) return;
+
+        log.trace("{} catchup {} - fetching chunk {}", req, phase, ++chunkCount);
+        List<PgFact> facts = jdbc.query(chunkQuery, extractor);
+        rowsToProcess = facts.size();
+        log.trace(
+            "{} catchup {} - processing chunk {} - found {} rows",
+            req,
+            phase,
+            chunkCount,
+            rowsToProcess);
+
+        // process them
+        facts.forEach(f -> pipeline.process(Signal.of(f)));
+      }
+      log.trace("{} catchup {} - all chunks processed", req, phase);
+    } else {
+      log.trace("{} catchup {} - no matching serials found", req, phase);
+    }
+
+    // tmp table is not needed anymore. As we reuse the connection, it'd be good to drop it.
     try {
-      // this needs to be transactional for fetch-size to have any effect whatsoever.
-      ds.getConnection().setAutoCommit(false);
-      jdbc.setFetchSize(props.getPageSize());
-      jdbc.setQueryTimeout(0); // disable query timeout
-      if (prepareTemporaryTable(jdbc, tempTableName) > 0) {
-
-        final var extractor = new PgFactExtractor(serial);
-
-        String chunkQuery = prepareChunkQuery(tempTableName);
-
-        int chunkCount = 0;
-        int rowsToProcess = -1;
-        while (rowsToProcess != 0) {
-
-          if (statementHolder.wasCanceled()) {
-            log.trace("{} catchup {} - was cancelled", req, phase);
-            return;
-          }
-
-          log.trace("{} catchup {} - fetching chunk {}", req, phase, ++chunkCount);
-          List<PgFact> facts = jdbc.query(chunkQuery, extractor);
-          rowsToProcess = facts.size();
-          log.trace(
-              "{} catchup {} - processing chunk {} - found {} rows",
-              req,
-              phase,
-              chunkCount,
-              rowsToProcess);
-
-          // process them
-          facts.forEach(f -> pipeline.process(Signal.of(f)));
-        }
-        log.trace("{} catchup {} - all chunks processed", req, phase);
-      } else {
-        log.trace("{} catchup {} - no matching serials found", req, phase);
-      }
+      if (wasCancelled()) return;
+      jdbc.execute("drop table " + tempTableName);
+      markConnectionDone();
     } catch (Exception e) {
-      if (!statementHolder.wasCanceled()) statementHolder.cancel();
-      throw e;
-    } finally {
-      // tmp table is not needed anymore. As we reuse the connection, it'd be good to drop it.
-      try {
-        jdbc.execute("drop table " + tempTableName);
-      } catch (Exception e) {
-        log.warn("{} catchup {} - while dropping tmp table:", req, phase, e);
-      }
+      log.warn("{} catchup {} - while dropping tmp table:", req, phase, e);
     }
   }
 
@@ -194,5 +189,11 @@ public class PgChunkedCatchup extends AbstractPgCatchup {
           req,
           elapsed.toSeconds());
     }
+  }
+
+  @Override
+  @VisibleForTesting
+  protected boolean wasCancelled() {
+    return super.wasCancelled();
   }
 }
